ml_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.2

pydantic==2.8.2
pydantic-settings==2.4.0
pyyaml==6.0.2

numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2

mlflow==2.18.0
zenml==0.66.0

pyspark==3.5.1